requirements.txt
asgiref==3.3.1
Django==3.1.6
django-countries==7.1
django-phonenumber-field==5.0.0
gunicorn==20.0.4
phonenumbers==8.12.20
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
